Adds calimoto changes to the mapsforge map writer module
This commit raises the basic gradle level to make the project compile.
The gradle settings are changed to reflect only the bare minimum needed
within the other projects.
The map writer module is changed according to the old mapsforge project
used by calimoto. This changes mostly remove the housenumbers from the
written map.

- Removes release function
- Changed Byte layout, also removes housenumbers from binary